refactor: remove interfaceAdapters folder on Auth domain
