+++ src/Auth/Presentation/Requests/AuthRequest.ts
@@ -1,4 +1,4 @@
-import AuthPayload from '../../InterfaceAdapters/Payloads/AuthPayload';
+import AuthPayload from '../../Domain/Payloads/AuthPayload';
 import MainConfig from '../../../Config/mainConfig';
 import { IsString, IsEmail, Length } from 'class-validator';
